pyproject.toml
[tool.poetry]
name = "dj-metersphere"
version = "0.1.0"
description = ""
authors = ["richard <richard_wu0313@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
